#2795-11
What: Modify DeployJob.py to join paths of destinations
and out_file names to rename the out_file to .so
Why: The functionality of Deploying requires the
filename to be changed to .so

What: Modify the DeployJob.py to handle multiple regions
Why: In case the user would like to deploy the COBOL
program to multiple regions. Functionality
works identically to passing multiple volumes.